Move package resolution helpers to gb/
After 08933c81 ResolvePackage{WithTests} was moved to cmd/gb/main.
To reduce the amount of code in a main package, these functions have
been moved to gb.

This reduces the amount of business logic in the main package, but leaves
the uncomfortable wart that ResolvePackageWithTests is in gb/, not gb/test/.
However, as ctx.ResolvePackageWithTests was already present inside gb/, this
is a wash for now.
