+++ resolve.go
@@ -1,10 +1,10 @@
-package main
+package gb
 
 import (
 	"fmt"
 	"go/build"
+	"path/filepath"
 
-	"github.com/constabulary/gb"
 	"github.com/constabulary/gb/log"
 )
 
@@ -13,18 +13,18 @@ type Resolver interface {
 	// Srcdirs returns []string{ "$PROJECT/src", "$PROJECT/vendor/src" }
 	Srcdirs() []string
 
-	// ResolvePackage resolves the import path to a *gb.Package
-	ResolvePackage(path string) (*gb.Package, error)
+	// ResolvePackage resolves the import path to a *Package
+	ResolvePackage(path string) (*Package, error)
 
 	// ResolvePackagesWithTests is similar to ResolvePackages however
 	// it also loads the test and external test packages of args into
 	// the context.
-	ResolvePackageWithTests(path string) (*gb.Package, error)
+	ResolvePackageWithTests(path string) (*Package, error)
 }
 
 // ResolvePackages resolves import paths to packages.
-func ResolvePackages(r Resolver, paths ...string) ([]*gb.Package, error) {
-	var pkgs []*gb.Package
+func ResolvePackages(r Resolver, paths ...string) ([]*Package, error) {
+	var pkgs []*Package
 	for _, path := range paths {
 		if path == "." {
 			return nil, fmt.Errorf("%q is not a package", r.Srcdirs()[0])
@@ -42,8 +42,8 @@ func ResolvePackages(r Resolver, paths ...string) ([]*gb.Package, error) {
 // ResolvePackagesWithTests is similar to ResolvePackages however
 // it also loads the test and external test packages of args into
 // the context.
-func ResolvePackagesWithTests(r Resolver, paths ...string) ([]*gb.Package, error) {
-	var pkgs []*gb.Package
+func ResolvePackagesWithTests(r Resolver, paths ...string) ([]*Package, error) {
+	var pkgs []*Package
 	for _, path := range paths {
 		path = relImportPath(r.Srcdirs()[0], path)
 		pkg, err := r.ResolvePackageWithTests(path)
@@ -57,4 +57,21 @@ func ResolvePackagesWithTests(r Resolver, paths ...string) ([]*gb.Package, error
 		pkgs = append(pkgs, pkg)
 	}
 	return pkgs, nil
+}
+
+func relImportPath(root, path string) string {
+	if isRel(path) {
+		var err error
+		path, err = filepath.Rel(root, path)
+		if err != nil {
+			log.Fatalf("could not convert relative path %q to absolute: %v", path, err)
+		}
+	}
+	return path
+}
+
+// isRel returns if an import path is relative or absolute.
+func isRel(path string) bool {
+	// TODO(dfc) should this be strings.StartsWith(".")
+	return path == "."
 }
